feat: add OTEL_SERVICE_NAME environment variable to cron jobs for enhanced observability
- Introduced the `OTEL_SERVICE_NAME` variable in multiple cron job configurations to improve observability and traceability in data ingestion processes.
- Each cron job now has a unique service name, aligning with the latest development standards and enhancing monitoring capabilities.

These updates ensure better tracking of data ingestion activities and improve overall system observability.

+++ src/dk_data/sqlmesh/models/cms/gold/cms_drug_market_profile.sql
@@ -45,7 +45,7 @@ SELECT
         WHEN ts.grand_total_spending > 0
         THEN ROUND(
             (d.total_part_d_spending + d.total_part_b_spending)::NUMERIC /
-            ts.grand_total_spending,
+            ts.grand_total_spending::NUMERIC,
             6
         )
         ELSE NULL
@@ -60,7 +60,7 @@ SELECT
         WHEN (d.total_part_d_claims + d.total_part_b_claims) > 0
         THEN ROUND(
             (d.total_part_d_spending + d.total_part_b_spending)::NUMERIC /
-            (d.total_part_d_claims + d.total_part_b_claims),
+            (d.total_part_d_claims + d.total_part_b_claims)::NUMERIC,
             2
         )
         ELSE NULL
